% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_code.R
\name{est_cfar}
\alias{est_cfar}
\title{Estimation of a CFAR Process}
\usage{
est_cfar(f, p = 3, df_b = 10, grid = 1000)
}
\arguments{
\item{f}{the functional time series.}

\item{p}{CFAR order.}

\item{df_b}{the degrees of freedom for natural cubic splines. Default is 10.}

\item{grid}{the number of gird points used to constrct the functional time series and noise process. Default is 1000.}
}
\value{
The function returns a list with components:
\item{phi_coef}{estimated spline coefficients for convluaional function(s).}
\item{phi_func}{estimated convoluational function(s).}
\item{rho}{estimated rho for O-U process (noise process).}
\item{sigma}{estimated sigma for O-U process (noise process).}
}
\description{
Estimation of a CFAR process.
}
\examples{
phi_func= function(x)   
{
 	return(dnorm(x,mean=0,sd=0.1))
}
library(MASS)	
library(splines)
grid=1000
y=g_cfar1(grid,5,phi_func)
f_grid=y$cfar
index=seq(1,grid+1,by=10)
f=f_grid[,index]
est=est_cfar(f,1)
b_grid=seq(-1,1,by=1/grid)
par(mfcol=c(1,1))
c1 <- range(est$phi_func)
plot(b_grid,phi_func(b_grid),type='l',col='red',ylim=c1*1.1)
points(b_grid,est$phi_func,type='l')
}
\references{
Liu, X., Xiao, H., and Chen, R. (2016) Convolutional autoregressive models for functional time series. \emph{Journal of Econometrics}, 194, 263-282.
}
