% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_code.R
\name{g_cfar}
\alias{g_cfar}
\title{Generate a CFAR Process}
\usage{
g_cfar(tmax = 1001, rho = 5, phi_list = NULL, grid = 1000, sigma = 1)
}
\arguments{
\item{tmax}{length of time.}

\item{rho}{parameter for O-U process (noise process).}

\item{phi_list}{the convolutional function(s). Default is the density function of normal distribution with mean 0 and standard deviation 0.1.}

\item{grid}{the number of grid points used to constrcut the functional time series. Default is 1000.}

\item{sigma}{the standard deviation of O-U process. Default is 1.}
}
\value{
The function returns a list with components:
\item{cfar}{a tmax-by-(grid+1) matrix following a CFAR(1) process.}
\item{epsilon}{the innovation at time tmax.}
}
\description{
Generate a convolutional functional autoregressive process.
}
\examples{
phi_func= function(x)   
{
return(dnorm(x,mean=0,sd=0.1))
}
library(MASS)	
library(splines)
test=g_cfar(1000,5,phi_func)

phi_func1= function(x){
return(0.5*x^2+0.5*x+0.13)
}
phi_func2= function(x){
return(0.7*x^4-0.1*x^3-0.15*x)
}
phi_list=list(phi_func1,phi_func2)
y=g_cfar(1000,5,phi_list)
}
\references{
Liu, X., Xiao, H., and Chen, R. (2016) Convolutional autoregressive models for functional time series. \emph{Journal of Econometrics}, 194, 263-282.
}
