% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_code.R
\name{Sstep.Clutter.Full.RB}
\alias{Sstep.Clutter.Full.RB}
\title{Sequential Importance Sampling under Clutter Environment}
\usage{
Sstep.Clutter.Full.RB(mm, xx, logww, yyy, par, xdim, ydim, resample.sch)
}
\arguments{
\item{mm}{the Monte Carlo sample size \code{m}.}

\item{xx}{the samples in the last iteration.}

\item{logww}{the log weight in the last iteration.}

\item{yyy}{the observations.}

\item{par}{a list of parameter values \code{(ssw,ssv,pd,nyy,yr)}, where \code{ssw} is the standard deviation in the state equation,
\code{ssv} is the standard deviation for the observation noise, \code{pd} is the probability to observe the true signal, \code{nyy} the dimension of the data,
and \code{yr} is the range of the data.}

\item{xdim}{the dimension of the state varible \code{x_t}.}

\item{ydim}{the dimension of the observation \code{y_t}.}

\item{resample.sch}{a binary vector of length \code{obs}, reflecting the resampling schedule. resample.sch[i]= 1 indicating resample should be carried out at step \code{i}.}
}
\value{
The function returns a list with the following components:
\item{xx}{the new sample.}
\item{logww}{the log weights.}
\item{xhat}{the fitted vlaues.}
\item{xhatRB}{the fitted values using Rao-Blackwellization.}
}
\description{
This function performs one step propagation using the sequential importantce sampling with full information proposal distribuiton and returns Rao-Blackwellization estimate of mean under clutter environment.
}
\examples{
nobs <- 100; pd <- 0.95; ssw <- 0.15; ssv <- 0.02;
xx0 <- 0; ss0 <- 0.1; nyy <- 10; yrange <- c(-80,80);
xdim <- 2; ydim <- nyy; nk <- 5
yr <- yrange[2]-yrange[1]
par <- list(ssw=ssw,ssv=ssv,nyy=nyy,pd=pd,yr=yr)
Yy10 <- array(dim=c(nyy,nobs,nk))
Xx10 <- matrix(ncol=nk,nrow=nobs)
Ii10 <- matrix(ncol=nk,nrow=nobs)
kk <- 0
seed.k <- 1
while(kk < nk){
	seed.k <- seed.k+1
	set.seed(seed.k)
	kk <- kk+1
	simu <- simuTargetClutter(nobs,pd,ssw,ssv,xx0,ss0,nyy,yrange)
	if(max(abs(simu$yy))>80){
		kk <- kk-1
	}else{
		Xx10[,kk] <- simu$xx; Yy10[,,kk] <- simu$yy; Ii10[,kk] <- simu$ii
	}
}
resample.sch <- rep.int(1,nobs)
delay <- 0; mm <- 2000;
Xxhat.p10 <- array(dim=c(nobs,1,nk))
Xxhat.RB10 <- array(dim=c(nobs,2,nk))
Xres.com1.10 <- array(dim=c(nobs,3,nk))
Xres.com2.10 <- array(dim=c(nobs,4,nk))
for(kk in 1:nk){
	xx.init <- matrix(nrow=2,ncol=mm)
	xx.init[1,] <- yrange[1]+runif(mm)*yr
	xx.init[2,] <- rep(0.1,mm)
	out3 <- SMC.Full.RB(Sstep.Clutter.Full.RB,nobs,Yy10[,,kk],mm,par,
	xx.init,xdim,ydim,resample.sch)
	Xxhat.RB10[,1,kk] <- out3$xhatRB[1,]
	Xxhat.RB10[,2,kk] <- out3$xhat[1,]
}
}
\references{
Tsay, R. and Chen, R. (2018) Nonlinear Time Series Analysis.
}
